Streamline configure and build process
jlebon on Freenode/#fedora-coreos had the wonderful
suggestion of using envsubst to provide certain tokens
to the environment in a preprocessing stage.
This is really convenient, fortunately, as envsubst will
only substitute variables that are in uppercase:
$UPPER or ${UPPER}. This makes some basic templating very
nice and easy.

Thus, ./configure is run first and performs envsubst on
_fcct-containers.yaml to produce fcct-containers.yaml.
That's ready for fcct to provide the Ignition JSON.

Additionally, ./build-znc is built from _build-znc and
envsubst generates ./run from _run. The former builds the
ZNC container and the run script launches FCOS with ZNC
container on Google Cloud Platform.

One thing currently missing: key-id (GPG) for the container
registry. That'll be a fixup. Unfortunately the registry
issue is the biggest downside of the current setup.
I did not find any official-looking ZNC container images,
tho it appears some people have made them on quay.io I have
not found any documentation.
Anyway, currently another VM hosts a registry container,
which is required in order to initially start ZNC when
booting FCOS - at least it likely will not be needed often.

On the off-chance anyone else wants to use these scripts I
it might be nice to have a README explaining some of the
configuration aspects.

That sums it up. envsubst allowed the removal of huge
chunks of base64-encoded public key data to be removed -
the ZNC configuration is still largely embedded in the YAML
template, but we give the container a persistent volume.
Therefore, rather than add templating and processing steps
for the ZNC config, just provision a minimal config.

Signed-of-by: Joanna