Use substenv and GPG keyservers to our advantage
Add lots of nice environment/CLI options to control the build
and deployment process. Discovered that when sigining docker
images, altho the error message sometimes says failed to write
external key/failed to sign it, it seems the problem is only a
local. It strikes when running rootless podman with no write
access to /var/lib/containers/sigstore.

Anyway after testing things a bit it seems to run fine including
both ./configure and ./build, not tried ./run yet. Had issues
earlier with whitespace in the public PGP key embedded in the
YAML template. Thus the switch to PGP key servers. For a moment
I thought I'd have to run my own, :^) but there are plenty of
public servers.

Also remove a lot of ZNC configuration from the _fcct-containers
YAML file. I think it's enough to start with a relatively minimal
configuration and add stuff later.

Signed-of-by: Joanna