Troubleshooting: make SSHD-only YAML template
Having fixed a few minor issues that gave obvious errors in
the instance console output, ZNC on FCOS still did not boot.
As the configuration was already rather complex, create a simple
template that provisions only SSH and disables docker - I have
noticed on the FAQ a warning that docker can cause unwanted problems,
even if you don't intentionally use it and are using podman.

Joanna

p.s. decided to remove default YAML filename and require that
the user supply one with --yml _file.yaml or file.yaml, envsubst
will only be executed if the underscore is not present